Feature/add group creation support (#11)
* Added group creation/lookup support
Updated docs and acceptance test for this new functionality
Add logic for erroring on empty required values
Added additional acceptance tests for multiple resources and datasource + resource scenarios

* Fix tfdocs for groups

* Add more tests for adding projects to groups dynamically

* Add group tests and dynamic lookups

* Add additional tests for group rights

* Update tests for groups and projects

* Add new examples for project and group

* Update libraries to latest version

* Fix examples and docs for project

* Fix comments and testing logic for updated role blocks

* Update group example

* update errors and tests

---------

+++ internal/resources/resource_cluster.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"strings"
 
-	paralusUtils "github.com/iherbllc/terraform-provider-paralus/internal/utils"
+	"github.com/iherbllc/terraform-provider-paralus/internal/utils"
 
 	"github.com/paralus/cli/pkg/cluster"
 	"github.com/paralus/cli/pkg/config"
@@ -154,7 +154,7 @@ func ResourceCluster() *schema.Resource {
 // Create a new cluster in Paralus
 func resourceClusterCreate(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
 
-	tflog.Debug(ctx, fmt.Sprintf("Provider Config Used: %s", paralusUtils.GetConfigAsMap(config.GetConfig())))
+	tflog.Debug(ctx, fmt.Sprintf("Provider Config Used: %s", utils.GetConfigAsMap(config.GetConfig())))
 
 	diags := createOrUpdateCluster(ctx, d, "POST")
 	if diags.HasError() {
@@ -168,7 +168,7 @@ func resourceClusterCreate(ctx context.Context, d *schema.ResourceData, m interf
 
 // Updating existing cluster
 func resourceClusterUpdate(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
-	tflog.Debug(ctx, fmt.Sprintf("Provider Config Used: %s", paralusUtils.GetConfigAsMap(config.GetConfig())))
+	tflog.Debug(ctx, fmt.Sprintf("Provider Config Used: %s", utils.GetConfigAsMap(config.GetConfig())))
 
 	return createOrUpdateCluster(ctx, d, "PUT")
 }
@@ -179,12 +179,22 @@ func createOrUpdateCluster(ctx context.Context, d *schema.ResourceData, requestT
 	projectId := d.Get("project").(string)
 	clusterId := d.Get("name").(string)
 
+	diags := utils.AssertStringNotEmpty("cluster project", projectId)
+	if diags.HasError() {
+		return diags
+	}
+
+	diags = utils.AssertStringNotEmpty("cluster name", clusterId)
+	if diags.HasError() {
+		return diags
+	}
+
 	tflog.Trace(ctx, fmt.Sprintf("Checking for project %s existance", projectId))
 
 	projectStruct, err := project.GetProjectByName(projectId)
 	if projectStruct == nil {
 		return diag.FromErr(errors.Wrap(err,
-			fmt.Sprintf("Project %s does not exist", projectId)))
+			fmt.Sprintf("project %s does not exist", projectId)))
 	}
 
 	howFail := "create"
@@ -192,7 +202,7 @@ func createOrUpdateCluster(ctx context.Context, d *schema.ResourceData, requestT
 		howFail = "update"
 	}
 
-	clusterStruct := paralusUtils.BuildClusterStructFromResource(d)
+	clusterStruct := utils.BuildClusterStructFromResource(d)
 
 	tflog.Trace(ctx, fmt.Sprintf("Cluster %s request", requestType), map[string]interface{}{
 		"cluster": clusterId,
@@ -200,22 +210,30 @@ func createOrUpdateCluster(ctx context.Context, d *schema.ResourceData, requestT
 	})
 
 	if requestType == "POST" {
+		// due to error swallowing, have to make sure the cluster doesn't exist before
+		// attempting to create it.
+		lookupStruct, _ := cluster.GetCluster(clusterId, projectId)
+		if lookupStruct != nil {
+			return diag.FromErr(errors.Wrap(err,
+				fmt.Sprintf("cluster %s already exists", clusterId)))
+		}
+
 		err := cluster.CreateCluster(clusterStruct)
 		if err != nil {
 			return diag.FromErr(errors.Wrap(err,
-				fmt.Sprintf("Failed to %s cluster %s in project %s", howFail,
+				fmt.Sprintf("failed to %s cluster %s in project %s", howFail,
 					clusterId, projectId)))
 		}
 	} else if requestType == "PUT" {
 		err := cluster.UpdateCluster(clusterStruct)
 		if err != nil {
 			return diag.FromErr(errors.Wrap(err,
-				fmt.Sprintf("Failed to %s cluster %s in project %s", howFail,
+				fmt.Sprintf("failed to %s cluster %s in project %s", howFail,
 					clusterId, projectId)))
 		}
 	} else {
 		return diag.FromErr(errors.Wrap(err,
-			fmt.Sprintf("Unknown request type %s", requestType)))
+			fmt.Sprintf("unknown request type %s", requestType)))
 	}
 
 	return resourceClusterRead(ctx, d, nil)
@@ -225,10 +243,20 @@ func createOrUpdateCluster(ctx context.Context, d *schema.ResourceData, requestT
 func resourceClusterRead(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
 	var diags diag.Diagnostics
 
-	tflog.Debug(ctx, fmt.Sprintf("Provider Config Used: %s", paralusUtils.GetConfigAsMap(config.GetConfig())))
+	tflog.Debug(ctx, fmt.Sprintf("Provider Config Used: %s", utils.GetConfigAsMap(config.GetConfig())))
 
 	projectId := d.Get("project").(string)
 	clusterId := d.Get("name").(string)
+
+	diags = utils.AssertStringNotEmpty("cluster project", projectId)
+	if diags.HasError() {
+		return diags
+	}
+
+	diags = utils.AssertStringNotEmpty("cluster name", clusterId)
+	if diags.HasError() {
+		return diags
+	}
 
 	tflog.Trace(ctx, "Retrieving cluster info", map[string]interface{}{
 		"cluster": clusterId,
@@ -245,9 +273,9 @@ func resourceClusterRead(ctx context.Context, d *schema.ResourceData, m interfac
 	}
 
 	// Update resource information from created/updated cluster
-	paralusUtils.BuildResourceFromClusterStruct(clusterStruct, d)
+	utils.BuildResourceFromClusterStruct(clusterStruct, d)
 
-	err = paralusUtils.SetBootstrapFileAndRelays(ctx, d)
+	err = utils.SetBootstrapFileAndRelays(ctx, d)
 	if err != nil {
 		return diag.FromErr(err)
 	}
@@ -258,13 +286,13 @@ func resourceClusterRead(ctx context.Context, d *schema.ResourceData, m interfac
 // Import cluster info into TF
 func resourceClusterImport(ctx context.Context, d *schema.ResourceData, m interface{}) ([]*schema.ResourceData, error) {
 
-	tflog.Debug(ctx, fmt.Sprintf("Provider Config Used: %s", paralusUtils.GetConfigAsMap(config.GetConfig())))
+	tflog.Debug(ctx, fmt.Sprintf("Provider Config Used: %s", utils.GetConfigAsMap(config.GetConfig())))
 
 	clusterProjectId := strings.Split(d.Id(), ":")
 
 	if len(clusterProjectId) != 2 {
 		d.SetId("")
-		return nil, errors.Wrap(nil, fmt.Sprintf("Unable to import. ID must be in format PROJECT_NAME:CLUSTER_NAME. Got %s", d.Id()))
+		return nil, errors.Wrap(nil, fmt.Sprintf("unable to import. ID must be in format PROJECT_NAME:CLUSTER_NAME. Got %s", d.Id()))
 	}
 
 	tflog.Trace(ctx, "Retrieving cluster info", map[string]interface{}{
@@ -277,12 +305,12 @@ func resourceClusterImport(ctx context.Context, d *schema.ResourceData, m interf
 	if err != nil {
 		d.SetId("")
 		// unlike others, we want to throw an error if the cluster does not exist so we can fail the import
-		return nil, errors.Wrap(err, fmt.Sprintf("Cluster %s does not exist in project %s",
+		return nil, errors.Wrap(err, fmt.Sprintf("cluster %s does not exist in project %s",
 			clusterProjectId[1], clusterProjectId[0]))
 	}
 
-	paralusUtils.BuildResourceFromClusterStruct(clusterStruct, d)
-	err = paralusUtils.SetBootstrapFileAndRelays(ctx, d)
+	utils.BuildResourceFromClusterStruct(clusterStruct, d)
+	err = utils.SetBootstrapFileAndRelays(ctx, d)
 	if err != nil {
 		d.SetId("")
 		return nil, err
@@ -298,26 +326,34 @@ func resourceClusterImport(ctx context.Context, d *schema.ResourceData, m interf
 func resourceClusterDelete(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
 	var diags diag.Diagnostics
 
-	tflog.Debug(ctx, fmt.Sprintf("Provider Config Used: %s", paralusUtils.GetConfigAsMap(config.GetConfig())))
+	tflog.Debug(ctx, fmt.Sprintf("Provider Config Used: %s", utils.GetConfigAsMap(config.GetConfig())))
 
 	projectId := d.Get("project").(string)
 	clusterId := d.Get("name").(string)
+
+	diags = utils.AssertStringNotEmpty("cluster project", projectId)
+	if diags.HasError() {
+		return diags
+	}
+
+	diags = utils.AssertStringNotEmpty("cluster name", clusterId)
+	if diags.HasError() {
+		return diags
+	}
 
 	tflog.Trace(ctx, "Deleting cluster info", map[string]interface{}{
 		"cluster": clusterId,
 		"project": projectId,
 	})
 
-	err := cluster.DeleteCluster(clusterId, projectId)
+	clusterStruct, _ := cluster.GetCluster(clusterId, projectId)
+	if clusterStruct != nil {
+		err := cluster.DeleteCluster(clusterId, projectId)
 
-	if err != nil {
-		// assume a no rows found error means the project does not exist
-		if strings.Contains(fmt.Sprint(err), "no rows in result set") {
-			d.SetId("")
-			return diags
+		if err != nil {
+			return diag.FromErr(errors.Wrap(err, fmt.Sprintf("failed to delete cluster %s in project %s",
+				clusterId, projectId)))
 		}
-		return diag.FromErr(errors.Wrap(err, fmt.Sprintf("Failed to delete cluster %s in project %s",
-			clusterId, projectId)))
 	}
 
 	d.SetId("")
